#24424 include in 23.01.5

+++ dotCMS/src/main/webapp/html/portlet/ext/roleadmin/view_roles_js_inc.jsp
@@ -114,15 +114,18 @@
 
 		// if tree is null, we are not filtering, so load the root nodes only calling RoleResource
 		if(tree==null) {
-			store = new dojox.data.JsonRestStore({ target: "/api/role/loadchildren/time/"+new Date()+"/id/", labelAttribute:"name", urlPreventCache: true});
+			treeStore = new dojox.data.JsonRestStore({
+				target: "/api/role/loadchildren/time/"+new Date()+"/id/",
+				labelAttribute:"name",
+				urlPreventCache: true});
 		} else { // if tree is not null, we need to build a store with the JSON tree contained in it
-			store = new dojo.data.ItemFileReadStore({ data: tree });
+			treeStore = new dojo.data.ItemFileReadStore({ data: tree });
 			autoExpand = true;
 		}
 
 
 	    treeModel = new dijit.tree.TreeStoreModel({
-	        store: store,
+	        store: treeStore,
 	        query: { top:true },
 	        rootId: "root",
 	        rootLabel: "Root",
@@ -164,13 +167,21 @@
 						newIcon.setAttribute("id", "lockedNode-" + item.id);
 						newIcon.classList = "lockIcon";
 						node.iconNode.parentNode.insertBefore(newIcon, node.iconNode);
+					} else if (!locked && dojo.byId("lockedNode-" + item.id)) {
+						var node = dijit.byId("treeNode-" + item.id);
+						node.iconNode.style.width = "0";
+						node.iconNode.style.height = "0";
+						node.iconNode.style.margin = "";
+						node.iconNode.classList.add("dijitInline");
+						node.iconNode.classList.add("dijitIcon");
+						node.iconNode.classList.add("dijitTreeIcon");
+						dojo.destroy("lockedNode-" + item.id);
 					}
 				}
 				return "";
 			},
 
 			getIconStyle: function (item, opened) {
-				var icon = "";
 				if(item) {
 					var locked = eval(norm(item.locked));
 					if(locked) {
@@ -190,26 +201,9 @@
 					}
 					var roleClickedDeferred = dojo.hitch(this, roleClicked, item.id);
 					roleClickedDeferredHandler = setTimeout(roleClickedDeferred, 500);
-
-				} else {
-					dijit.byId('editRoleButtonWrapper').setAttribute("disabled", true);
-					dijit.byId('deleteRoleButtonWrapper').setAttribute("disabled", true);
-					dojo.style(dojo.byId('roleTabs'), { display: 'none' });
-					currentRoleId=null;
-					currentRole=null;
-				}
-
-				// used to remove the highlight (bold) from the last selected treeNode
-				if(lastSelectedNode && lastSelectedNode.labelNode) {
-					lastSelectedNode.labelNode.style.fontWeight="normal";
-					lastSelectedNode = null;
-				}
-
-				// used to highlight (bold) the current selected treeNode
-				if(dijit.byId("rolesTree").selectedNode) {
-					dijit.byId("rolesTree").selectedNode.labelNode.style.fontWeight="bold";
 					lastSelectedNode = dijit.byId("rolesTree").selectedNode;
-
+				} else {
+					clearSelectedRole();
 				}
 
 			},
@@ -246,12 +240,30 @@
         // when we right-click anywhere on the tree, make sure we open the menu
         menu.bindDomNode(dojo.byId('rolesTree'));
 
+		dojo.connect(tree, "onMouseDown", this, (function(e) {
+			if (e.button == "2") {
+				var selectedNode = dijit.byId("rolesTree").selectedNode;
+				var selectedNodeId = selectedNode && selectedNode.item ? selectedNode.item.id : null;
+				var lastSelectedNodeId = lastSelectedNode && lastSelectedNode.item ? lastSelectedNode.item.id : null;
+				if (!lastSelectedNodeId || selectedNodeId != lastSelectedNodeId) {
+					setSelectedRoleInTree();
+				}
+			}
+		}).bind(this));
+
 		dojo.connect(menu, "_openMyself", this, (function(e) {
 
             var tn = dijit.getEnclosingWidget(e.target);
-			var item = tn.lastFocused;
 
-			var role = findRole(item.id.replace("treeNode-",""));
+			var item = tn.item && tn.item.id ? tn.item :
+					tree.lastFocused && tree.lastFocused.item ?
+							tree.lastFocused.item : null;
+			var role = item && item.id ? (item.id.includes("treeNode") ? findRole(item.id.replace("treeNode-", "")) : findRole(item.id)) : null;
+			if (!role && item && item.id) {
+				console.info("No role found for item: " + item.id);
+				return;
+			}
+
 			var locked = eval(norm(role.locked));
 			var system = eval(norm(role.system));
 
@@ -289,6 +301,10 @@
 			deleteRoleMenuRefHandle = dojo.connect(dojo.byId('deleteRoleMenu'), 'onclick', this, dojo.hitch(this, deleteRole.bind(this), norm(role.id)));
 
         }).bind(this));
+
+		if (lastSelectedNode) {
+			setSelectedRoleInTree();
+		}
 	}
 
 	var editRoleMenuRefHandle;
@@ -310,8 +326,6 @@
 
 	//The logic is executed deferred within 50 ms to handle multiple user keystrokes
 	function filterRolesDeferred () {
-		var tree = dijit.byId('rolesTree');
-		var matchesCount = 0;
 
 		dojo.style(dojo.byId('noRolesFound'), { display: 'none' });
 
@@ -324,16 +338,11 @@
 			setTimeout("hideToolTip()", 3000);
 		} else if(name=='') {// if filter textbox is empty
 			hideToolTip();
-			lastSelectedNode = null;
+			clearSelectedRole();
 			buildRolesTree(null);
-			dijit.byId('editRoleButtonWrapper').setAttribute("disabled", true);
-			dijit.byId('deleteRoleButtonWrapper').setAttribute("disabled", true);
-			dojo.style(dojo.byId('roleTabs'), { display: 'none' });
-			currentRoleId=null;
-			currentRole=null;
 		} else if(name!=null && name.length>2) {// if filter textbox has at least 3 chars, we build the resulting tree
 			hideToolTip();
-			lastSelectedNode = null;
+			clearSelectedRole();
 			var filteredRoles = searchRoles(name);
 			buildRolesTree(filteredRoles);
 
@@ -362,24 +371,38 @@
 		filterRolesDeferred()
 	}
 
+	function setSelectedRoleInTree() {
+		var path = [];
+		if (lastSelectedNode) {
+			var treePath = lastSelectedNode.getTreePath();
+			for (var i = 0; i < treePath.length; i++) {
+				var item = treePath[i];
+				path.push(item.id);
+			}
+		}
+		var tree = dijit.byId('rolesTree');
+		tree.set("path", path);
+	}
+
 	//CRUD operations over roles
 
 	//Executed when a grid row is clicked
-	var newRole = false;
+	var isNewRole = false;
+	var roleToSaveId = null;
 	function editRole(roleId) {
 
 		if(!roleId) {
-			roleId = currentRoleId;
+			roleId = norm(currentRoleId);
 		}
 
 
 		var role = findRole(roleId);
 
-		currentRoleId = role.id;
-		currentRole = role;
-
+		if(!role) {
+			return;
+		}
 		isNewRole = false;
-		setRoleName(role);
+		roleToSaveId = roleId;
 
 		dijit.byId('roleName').attr('value', norm(role.name));
 		dijit.byId('roleKey').attr('value', norm(role.roleKey));
@@ -387,8 +410,8 @@
 
 		var parent = findRole(parentRoleId);
 // 		dijit.byId('parentRole').set('value', parent.id==role.id?"Root Role":parent.name);
-		dojo.byId('parentRoleDiv').innerHTML = parent.id==role.id?"Root Role":parent.name;
-		dojo.byId('parentRoleValue').value = parent.id==role.id?"0":parent.id
+		dojo.byId('parentRoleDiv').innerHTML = parent && parent.id==role.id?"Root Role":parent.name;
+		dojo.byId('parentRoleValue').value = parent && parent.id==role.id?"0":parent.id
 // 		dijit.byId('parentRole').set('displayedValue', parent.id==role.id?"Root Role":parent.name);
 		dijit.byId('editUsers').attr('value', norm(role.editUsers) == true?true:false);
 		dijit.byId('editPermissions').attr('value', norm(role.editPermissions) == true?true:false);
@@ -401,10 +424,10 @@
 	//Executed when adding a new user
 
 
-	var isNewRole = false;
 	function addNewRole() {
 
 		isNewRole = true;
+		roleToSaveId = null;
 
 		dojo.byId('addRoleErrorMessagesList').innerHTML = '';
 		dijit.byId('roleName').reset();
@@ -437,6 +460,7 @@
 
 	//Handler when the user clicks the cancel button
 	function cancelAddNewRole () {
+		roleToSaveId = null;
 		dijit.byId('addRoleDialog').hide();
 	}
 
@@ -455,9 +479,6 @@
 		var canEditLayouts = dijit.byId('editTabs').checked;
 		var description = dijit.byId('roleDescription').attr('value');
 
-		console.log(parentRoleId);
-		currentRoleId = norm(currentRoleId);
-
 		if(isNewRole) {
 			RoleAjax.addNewRole(roleName, roleKey, (parentRoleId == 0?null:parentRoleId), canEditUsers, canEditPermissions, canEditLayouts, description, {
 				callback: saveRoleCallback,
@@ -465,7 +486,7 @@
 			});
 		}
 		else {
-			RoleAjax.updateRole(currentRoleId, roleName, roleKey, (parentRoleId == 0?null:parentRoleId), canEditUsers, canEditPermissions, canEditLayouts, description, {
+			RoleAjax.updateRole(roleToSaveId, roleName, roleKey, (parentRoleId == 0?null:parentRoleId), canEditUsers, canEditPermissions, canEditLayouts, description, {
 				callback: saveRoleCallback,
 				exceptionHandler: saveRoleExceptionHandler
 			});
@@ -473,15 +494,74 @@
 
 	}
 
-	function saveRoleCallback (newRole) {
-		var roleName = norm(newRole.name);
+	function saveRoleCallback (savedRole) {
 		dijit.byId('addRoleDialog').hide();
-		dojo.byId("displayRoleName1").innerHTML= roleName;
-		dojo.byId("displayRoleName2").innerHTML= roleName;
-		dojo.byId("displayRoleName3").innerHTML = roleName;
-		lastSelectedNode =  null;
-		buildRolesTree();
-
+		// 	setRoleName(currentRole);
+		setRoleName(savedRole);
+		var tree = dijit.byId('rolesTree');
+		if (isNewRole) {
+			// add new role to the tree if needed
+			tree.model.getRoot(function (rootItem) {
+				showNewRoleOnTree(rootItem, savedRole);
+			}, function (error) {
+				console.error("couldn't get root role", error);
+				clearSelectedRole();
+				buildRolesTree();
+			});
+		} else {
+			// update label for the saved role
+			var nodeList = tree.getNodesByItem(savedRole.id);
+			if (nodeList && nodeList.length > 0 && nodeList[0]) {
+				var node = nodeList[0];
+				if (node.label != savedRole.name) {
+					node.set("label",savedRole.name);
+				}
+			}
+		}
+	}
+	function showNewRoleOnTree(rootItem, savedRole) {
+		var parentItem = rootItem;
+		if (lastSelectedNode && lastSelectedNode.item) {
+			parentItem = lastSelectedNode.item;
+		}
+		var tree = dijit.byId('rolesTree');
+		tree.model.getChildren(parentItem, function (children) {
+			var newChildren = [];
+			var newChildAdded = false
+			for (var i = 0; i < children.length; i++) {
+				newChildren.push(children[i]);
+				if (children[i].id  == savedRole.id) {
+					newChildAdded = true;
+				}
+			}
+			if (!newChildAdded) {
+				var newRoleItem = {
+					"id": savedRole.id,
+					"$ref": savedRole.id,
+					"name": savedRole.name,
+					"locked": savedRole.locked,
+					"children": true
+				};
+				newChildren.push(newRoleItem);
+				var nodeList = tree.getNodesByItem(parentItem.id);
+				if (nodeList && nodeList.length > 0 && nodeList[0]) {
+					var parentNode = nodeList[0];
+					parentNode.setChildItems(newChildren);
+				}
+			}
+		}, function (error) {
+			console.error("couldn't get children for role", error);
+			clearSelectedRole();
+			buildRolesTree();
+		});
+	}
+	function clearSelectedRole() {
+		dojo.style(dojo.byId('roleTabs'), { display: 'none' });
+		dijit.byId('deleteRoleButtonWrapper').setAttribute("disabled", true);
+		dijit.byId('editRoleButtonWrapper').setAttribute("disabled", true);
+		lastSelectedNode = null;
+		currentRoleId=null;
+		currentRole=null;
 	}
 
 	function saveRoleExceptionHandler (message, exception) {
@@ -498,24 +578,47 @@
 	}
 
 	//Event handler then deleting a user
+	var roleToDeleteId = null;
 	function deleteRole(roleId) {
-		if(!roleId)
+		roleToDeleteId = null;
+		if(!roleId) {
 			roleId = norm(currentRoleId);
+		}
 
-		if(confirm(confirmRemoveRoleMsg))
+		if(roleId && confirm(confirmRemoveRoleMsg)) {
 			RoleAjax.deleteRole(roleId, deleteRoleCallback);
+			roleToDeleteId = roleId;
+		}
 	}
 
 	//Callback from the server to confirm a user deletion
 	function deleteRoleCallback (isDeleted) {
-		dojo.style(dojo.byId('roleTabs'), { display: 'none' });
-		dijit.byId('deleteRoleButtonWrapper').setAttribute("disabled", true);
-		dijit.byId('editRoleButtonWrapper').setAttribute("disabled", true);
-		lastSelectedNode = null;
-		currentRoleId=null;
-		currentRole=null;
-		buildRolesTree();
 		if(isDeleted){
+			var nodeIsInTreePath = false;
+			if (lastSelectedNode) {
+				var treePath = lastSelectedNode.getTreePath();
+				for (var i = 0; i < treePath.length; i++) {
+					var item = treePath[i];
+					if (item.id == roleToDeleteId) {
+						nodeIsInTreePath = true;
+						break;
+					}
+				}
+			}
+			if (!currentRoleId || nodeIsInTreePath) {
+				clearSelectedRole();
+				buildRolesTree();
+			} else {
+				var tree = dijit.byId('rolesTree');
+				var nodeList = tree.getNodesByItem(roleToDeleteId);
+				if (nodeList && nodeList.length > 0 && nodeList[0]) {
+					var node = nodeList[0];
+					var parent = node.getParent();
+					if (parent) {
+						parent.removeChild(node);
+					}
+				}
+			}
 			showDotCMSSystemMessage(roleRemovedMsg);
 		}else{
 			showDotCMSSystemMessage(roleNotRemovedMsg);
@@ -534,8 +637,16 @@
 	}
 
 	function lockRoleCallback (lockedRoleId) {
-		dijit.byId('editRoleButtonWrapper').setAttribute("disabled", norm(currentRoleId) == norm(lockedRoleId));
-		buildRolesTree();
+		if (currentRoleId) {
+			if (norm(currentRoleId) == norm(lockedRoleId)) {
+				currentRole.locked = true;
+			}
+			dijit.byId('editRoleButtonWrapper').setAttribute("disabled",
+					eval(norm(currentRole.system)) || eval(norm(currentRole.locked)));
+			dijit.byId('deleteRoleButtonWrapper').setAttribute("disabled",
+					eval(norm(currentRole.system)) || eval(norm(currentRole.locked)) || currentRole.children.length > 0);
+		}
+		lockOrUnlockRoleInTree(lockedRoleId);
 		showDotCMSSystemMessage(roleLockedMsg);
 
 	}
@@ -546,9 +657,29 @@
 		node.iconNode.style.width = '0px';
 		node.iconNode.style.height = '0px';
 
-		dijit.byId('editRoleButtonWrapper').setAttribute("disabled", norm(currentRoleId) == norm(unlockedRoleId) && !eval(norm(currentRole.system)));
-		buildRolesTree();
+		if (currentRoleId) {
+			if (norm(currentRoleId) == norm(unlockedRoleId)) {
+				currentRole.locked = false
+			}
+			dijit.byId('editRoleButtonWrapper').setAttribute("disabled",
+					eval(norm(currentRole.system)) || eval(norm(currentRole.locked)));
+			dijit.byId('deleteRoleButtonWrapper').setAttribute("disabled",
+					eval(norm(currentRole.system)) || eval(norm(currentRole.locked)) || currentRole.children.length > 0);
+		}
+		lockOrUnlockRoleInTree(unlockedRoleId);
 		showDotCMSSystemMessage(roleUnlockedMsg);
+	}
+
+	// this function receive an id and based in the id we lock or unlock the role
+	function lockOrUnlockRoleInTree (itemId) {
+		var tree = dijit.byId('rolesTree');
+		//we are using this function, because in some cases, the node list is empty
+		const itemProps = findRole(itemId);
+		//itemProps contains the required data to determine if the role is locked
+		if (itemProps) {
+			tree.getIconClass(itemProps, false);
+			tree.getIconStyle(itemProps, false);
+		}
 	}
 
 
@@ -562,6 +693,11 @@
 			dojo.byId("displayRoleName1").innerHTML= roleName;
 			dojo.byId("displayRoleName2").innerHTML= roleName;
 			dojo.byId("displayRoleName3").innerHTML = roleName;
+			// only when update
+			const updHtml = dojo.byId(`treeNode-${role.id}_label`)
+			if (updHtml !== null){
+				updHtml.innerHTML = roleName
+			}
 		}
 	}
 
@@ -577,13 +713,18 @@
 
 		currentRoleId = roleId;
 		var role = findRole(roleId);
+		if (!role) {
+			currentRoleId = null;
+			return;
+		}
 		currentRole = role;
 		setRoleName(role);
 		dojo.byId('roleKey').innerHTML = norm(role.roleKey);
 		dojo.byId('rolePath').innerHTML = norm(role.FQN);
 
 		dijit.byId('editRoleButtonWrapper').setAttribute("disabled", eval(norm(role.system)) || eval(norm(role.locked)));
-		dijit.byId('deleteRoleButtonWrapper').setAttribute("disabled", eval(norm(role.system)) || role.children.length > 0);
+		dijit.byId('deleteRoleButtonWrapper').setAttribute("disabled",
+				eval(norm(role.system)) || eval(norm(role.locked)) || role.children.length > 0);
 
 		renderCurrentTab();
 	}
@@ -1417,8 +1558,8 @@
 		while(parentId && parentId != roleId) {
 			role = findRole(role.parent);
 			branches.push(role);
-			var parentId = dojo.isArray(role.parent)?role.parent[0]:role.parent;
-			var roleId = dojo.isArray(role.id)?role.id[0]:role.id;
+			parentId = dojo.isArray(role.parent)?role.parent[0]:role.parent;
+			roleId = dojo.isArray(role.id)?role.id[0]:role.id;
 		}
 
 		return branches;
